s3 upload done, change in public and private route

+++ src/utils/aws/s3.aws.ts
@@ -1,6 +1,5 @@
 import { PutObjectCommand, S3Client } from "@aws-sdk/client-s3";
 import fs from 'fs';
-import { ObjectCannedACL } from "@aws-sdk/client-s3";
 
 const s3Client = new S3Client({
     region:process.env.AWS_REGION||'',
@@ -12,7 +11,7 @@ const s3Client = new S3Client({
 })
 
 
-export async function uploadImageToS3(file:any){
+export async function uploadImageToS3(file:any):Promise<string|null>{
     // upload image to 
     console.log(file)
     const uploadParams = {
@@ -20,13 +19,20 @@ export async function uploadImageToS3(file:any){
         Key:file.originalname,
         Body:fs.createReadStream(file.path),
         ContentType:file.mimetype,
-        ACL:'public-read' as ObjectCannedACL, // Cast the string value to ObjectCannedACL type
     }
 
-    const s3Response = await s3Client.send(new PutObjectCommand(uploadParams));
-    console.log("this is s3Response ", s3Response);
-    // delete the file from local storage
-    fs.unlinkSync(file.path);
-
+    try {
+        const s3Response = await s3Client.send(new PutObjectCommand(uploadParams));
+        console.log("this is s3Response ", s3Response);
+        // Construct the URL of the uploaded object
+        const objectUrl = `https://${uploadParams.Bucket}.s3.amazonaws.com/${uploadParams.Key}`;
+        // delete the file from local storage
+        if(fs.existsSync(file.path)){
+            fs.unlinkSync(file.path)
+        }
+        return objectUrl;
+    } catch (error) {
+        return null;
+    }
 
 }
